feat(command-palette): fix keyboard nav, unify action path, mirror overview panel source
Command palette keyboard + panel-list fixes:

- Fix arrow/Enter navigation: handle keys on the search input's own onKeyDown
  with autoFocus instead of a document-capture listener that lost the keys when
  the palette opened over a focused terminal/canvas. Make the selected-row
  highlight visible and scroll it into view as selection moves off-screen.
- Remove the per-row keybind display.
- Unify panel creation with the keyboard/menu path: extract runAction into
  lib/runAction so the palette, shortcuts, and native menu share one
  context-aware path (drops onto the focused canvas or tabs into the focused
  dock stack, with the ensure-folder prompt).
- Add New Cate Agent (Cmd+Shift+A) and New Canvas (Cmd+Shift+C) shortcuts and
  menu items; add the discrete actions missing from the palette (Close Panel,
  Save File, Toggle Search, Toggle Minimap, Zoom to Selection, Undo, Redo).
- Mirror the sidebar overview's panel source: extract useWorkspacePanelTree
  (ws.panels joined against every canvas store + the dock store, ghost-filtered)
  and use it in both WorkspaceTab and the palette, so they can't disagree about
  which panels exist or where they live.
