+++ src/renderer/hooks/useShortcuts.ts
@@ -11,15 +11,17 @@ import {
   getActiveCanvasOps,
   getActiveCanvasPanelId,
   getWorkspaceCanvasStore,
-  placementForActivePanel,
 } from '../stores/appStore'
-import { useUIStore, getSidebarLayout } from '../stores/uiStore'
-import { useSearchStore } from '../stores/searchStore'
+import { useUIStore } from '../stores/uiStore'
 import { getActivePanelId, setActivePanel } from '../lib/activePanel'
 import { resolvePanelById } from '../lib/workspace/panelReveal'
 import { getNodeActivePanelId } from '../panels/nodeDockRegistry'
-import type { MenuActionId, ShortcutAction } from '../../shared/types'
-import { confirmClosePanels } from '../lib/confirmClosePanels'
+import type { ShortcutAction } from '../../shared/types'
+import { runAction } from '../lib/runAction'
+
+// ensureWorkspaceFolder lives in lib/runAction now; re-exported here for the
+// panels/pages that still import it from this module.
+export { ensureWorkspaceFolder } from '../lib/runAction'
 
 // Cmd+Arrow panel navigation — moves the selection cursor between nodes.
 const NAVIGATE_ACTIONS = new Set<ShortcutAction>([
@@ -30,22 +32,6 @@ const NAVIGATE_ACTIONS = new Set<ShortcutAction>([
 const PAN_ACTIONS = new Set<ShortcutAction>([
   'panUp', 'panDown', 'panLeft', 'panRight',
 ])
-
-/**
- * Ensures the workspace has a rootPath before proceeding.
- * If no rootPath is set, opens the folder dialog first.
- * Returns the workspaceId if ready, or null if the user cancelled.
- */
-export async function ensureWorkspaceFolder(workspaceId: string): Promise<string | null> {
-  const ws = useAppStore.getState().getWorkspace(workspaceId)
-  if (ws?.rootPath) return workspaceId
-
-  const folderPath = await window.electronAPI.openFolderDialog()
-  if (!folderPath) return null
-
-  useAppStore.getState().setWorkspaceRootPath(workspaceId, folderPath)
-  return workspaceId
-}
 
 /**
  * Whether a terminal panel currently holds input focus, derived from the
@@ -113,178 +99,10 @@ export function useShortcuts(): void {
     const canvasStore = () => (getActiveCanvasOps()?.storeApi ?? canvasStoreApi).getState()
     const appStore = useAppStore.getState
 
-    /**
-     * Run a shortcut/menu action. Shared between the keyboard handler and the
-     * native menu IPC listener, so the two code paths can never drift.
-     * Re-reads store state at call time so it's safe to invoke at any moment.
-     */
-    async function runAction(action: MenuActionId): Promise<void> {
-      const selectedWorkspaceId = appStore().selectedWorkspaceId
-
-      // Menu-only actions first
-      if (action === 'openFolder') {
-        const folder = await window.electronAPI.openFolderDialog()
-        if (folder) {
-          useAppStore.getState().setWorkspaceRootPath(selectedWorkspaceId, folder)
-        }
-        return
-      }
-      if (action === 'reloadWorkspace') {
-        const { reloadActiveWorkspaceFromDisk } = await import('../lib/workspace/session')
-        await reloadActiveWorkspaceFromDisk()
-        return
-      }
-      if (action === 'manageLayouts') {
-        useUIStore.getState().setShowLayoutsDialog(true)
-        return
-      }
-
-      switch (action as ShortcutAction) {
-        case 'newTerminal': {
-          const placement = placementForActivePanel()
-          const wsId = await ensureWorkspaceFolder(selectedWorkspaceId)
-          if (wsId) appStore().createTerminal(wsId, undefined, undefined, placement)
-          break
-        }
-        case 'newBrowser': {
-          const placement = placementForActivePanel()
-          const wsId = await ensureWorkspaceFolder(selectedWorkspaceId)
-          if (wsId) appStore().createBrowser(wsId, undefined, undefined, placement)
-          break
-        }
-        case 'newEditor':
-        case 'newFile': {
-          const placement = placementForActivePanel()
-          const wsId = await ensureWorkspaceFolder(selectedWorkspaceId)
-          if (wsId) appStore().createEditor(wsId, undefined, undefined, placement)
-          break
-        }
-        case 'closePanel': {
-          const focusedNodeId = canvasStore().focusedNodeId
-          if (focusedNodeId) {
-            const node = canvasStore().nodes[focusedNodeId]
-            if (node && (await confirmClosePanels(selectedWorkspaceId, [node.panelId]))) {
-              appStore().closePanel(selectedWorkspaceId, node.panelId)
-            }
-          }
-          break
-        }
-        case 'toggleSidebar':
-          useUIStore.getState().toggleSidebar()
-          break
-        case 'toggleFileExplorer': {
-          const ui = useUIStore.getState()
-          const side = getSidebarLayout().left.includes('explorer') ? 'left' : 'right'
-          if (side === 'left') {
-            ui.setActiveLeftSidebarView(ui.activeLeftSidebarView === 'explorer' ? null : 'explorer')
-          } else {
-            ui.setActiveRightSidebarView(ui.activeRightSidebarView === 'explorer' ? null : 'explorer')
-          }
-          break
-        }
-        case 'toggleSearch': {
-          const ui = useUIStore.getState()
-          const side = getSidebarLayout().left.includes('search') ? 'left' : 'right'
-          const active = side === 'left' ? ui.activeLeftSidebarView : ui.activeRightSidebarView
-          const next = active === 'search' ? null : 'search'
-          if (side === 'left') ui.setActiveLeftSidebarView(next)
-          else ui.setActiveRightSidebarView(next)
-          if (next === 'search') useSearchStore.getState().requestFocus()
-          break
-        }
-        case 'toggleMinimap':
-          useUIStore.getState().toggleMinimapOpen()
-          break
-        case 'nodeSwitcher':
-          useUIStore.getState().setShowNodeSwitcher(true)
-          break
-        case 'commandPalette':
-          useUIStore.getState().setShowCommandPalette(true)
-          break
-        case 'zoomIn':
-          canvasStore().animateZoomTo(canvasStore().zoomLevel + 0.1)
-          break
-        case 'zoomOut':
-          canvasStore().animateZoomTo(canvasStore().zoomLevel - 0.1)
-          break
-        case 'zoomReset':
-          canvasStore().animateZoomTo(1.0)
-          break
-        case 'focusNext': {
-          const next = canvasStore().nextNode()
-          if (next) canvasStore().focusNode(next)
-          break
-        }
-        case 'focusPrevious': {
-          const prev = canvasStore().previousNode()
-          if (prev) canvasStore().focusNode(prev)
-          break
-        }
-        case 'saveFile':
-          window.dispatchEvent(new CustomEvent('save-file'))
-          break
-        case 'zoomToFit':
-          canvasStore().zoomToFit()
-          break
-        case 'zoomToSelection':
-          canvasStore().zoomToSelection()
-          break
-        case 'toolSelect':
-          useUIStore.getState().setActiveTool('select')
-          break
-        case 'toolHand':
-          useUIStore.getState().setActiveTool('hand')
-          break
-        case 'navigateUp':
-          canvasStore().navigateSelect('up')
-          break
-        case 'navigateDown':
-          canvasStore().navigateSelect('down')
-          break
-        case 'navigateLeft':
-          canvasStore().navigateSelect('left')
-          break
-        case 'navigateRight':
-          canvasStore().navigateSelect('right')
-          break
-        case 'panUp':
-          canvasStore().panViewport('up')
-          break
-        case 'panDown':
-          canvasStore().panViewport('down')
-          break
-        case 'panLeft':
-          canvasStore().panViewport('left')
-          break
-        case 'panRight':
-          canvasStore().panViewport('right')
-          break
-        case 'autoLayout':
-          canvasStore().autoLayout()
-          break
-        case 'undo':
-          canvasStore().undo()
-          break
-        case 'redo':
-          canvasStore().redo()
-          break
-        case 'deleteNode': {
-          const focusedId = canvasStore().focusedNodeId
-          if (focusedId && canvasStore().nodes[focusedId]) {
-            const node = canvasStore().nodes[focusedId]
-            if (await confirmClosePanels(selectedWorkspaceId, [node.panelId])) {
-              appStore().closePanel(selectedWorkspaceId, node.panelId)
-            }
-          }
-          break
-        }
-      }
-    }
-
     // Subscribe to native-menu dispatches. The menu fires this on every File /
     // View / Terminal / etc. item that maps to a runnable action.
     const unsubscribeMenu = window.electronAPI.onMenuTriggerAction((action) => {
-      runAction(action).catch(() => { /* noop — menu actions are best-effort */ })
+      runAction(action, canvasStoreApi).catch(() => { /* noop — menu actions are best-effort */ })
     })
 
     // A panel-creation shortcut fired while a detached dock/panel window was
@@ -294,7 +112,7 @@ export function useShortcuts(): void {
       if (workspaceId && appStore().getWorkspace(workspaceId) && appStore().selectedWorkspaceId !== workspaceId) {
         void appStore().selectWorkspace(workspaceId)
       }
-      runAction(action).catch(() => { /* noop */ })
+      runAction(action, canvasStoreApi).catch(() => { /* noop */ })
     })
 
     // Native "Layouts" menu → load a saved layout into the active canvas.
@@ -480,7 +298,7 @@ export function useShortcuts(): void {
       e.preventDefault()
       e.stopPropagation()
 
-      runAction(action).catch(() => { /* noop */ })
+      runAction(action, canvasStoreApi).catch(() => { /* noop */ })
     }
 
     function handleKeyUp(e: KeyboardEvent) {
